+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.1.0-alpha34] - 2024-09-30
+
+- improve perf: `tagger` will cache the loaded images in memory
+
 ## [0.1.0-alpha33] - 2024-09-30
 
 - improve perf and fix bug: `tagger pick`